pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "orcasync"
version = "0.2.1"
description = "Sync OrcaSlicer profiles between multiple machines using Git"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["orcaslicer", "sync", "3d-printing", "profiles"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "gitpython>=3.1.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "textual>=7.3.0",
]

[project.scripts]
orcasync = "orcasync.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["orcasync"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",
    "pytest-benchmark>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "textual-dev>=1.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pyinstaller>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/orcasync"
"Bug Tracker" = "https://github.com/yourusername/orcasync/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=orcasync",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "benchmark: marks tests as benchmark tests",
]
# pytest-benchmark settings
[tool.pytest.ini_options.benchmark]
max_time = "0.5"
min_rounds = 5
calibration_precision = 10
warmup = true
disable_gc = true
